refactor : popup 디렉토리 이동
